fix eslint errors

+++ modules/user/user.service.js
@@ -50,11 +50,10 @@ const {
   ONLY_SUPER_ADMIN_CAN_UNLOCK_ADMIN,
   ONLY_SUPER_ADMIN_CAN_BLOCK_ADMIN,
   INVALID_OTP_CODE,
-  TOKEN_IS_EXPIRIED,
 } = require('../../error-messages/user.messages');
 const FilterHelper = require('../../helpers/filter-helper');
 const {
-  STATUS_CODES: { NOT_FOUND, BAD_REQUEST, FORBIDDEN, UNAUTHORIZED },
+  STATUS_CODES: { NOT_FOUND, BAD_REQUEST, FORBIDDEN },
 } = require('../../consts/status-codes');
 const {
   USER_BLOCK_PERIOD: { UNLOCKED, ONE_MONTH, TWO_MONTH, INFINITE },
@@ -70,10 +69,13 @@ const {
   roles: { USER },
 } = require('../../consts');
 const RuleError = require('../../errors/rule.error');
-const { USER_IS_BLOCKED } = require('../../error-messages/user.messages');
+const {
+  USER_IS_BLOCKED,
+  SUPER_ADMIN_IS_IMMUTABLE,
+} = require('../../error-messages/user.messages');
 const {
   roles: { ADMIN, SUPERADMIN },
-} = require('../../consts/');
+} = require('../../consts');
 const {
   HISTORY_ACTIONS: {
     BLOCK_USER: BLOCK_USER_ACTION,
@@ -182,6 +184,8 @@ class UserService extends FilterHelper {
 
         break;
       }
+      default:
+        break;
     }
 
     const { beforeChanges, afterChanges } = getChanges(
@@ -310,15 +314,22 @@ class UserService extends FilterHelper {
       .populate('orders')
       .exec();
     const paidOrders = user.orders.filter(order => order.isPaid);
-    return paidOrders.reduce((acc, order) => {
-      acc = [...acc, ...order.items.map(item => ({ _id: item.productId }))];
-      return acc;
-    }, []);
+    return paidOrders.reduce(
+      (acc, order) => [
+        ...acc,
+        ...order.items.map(item => ({ _id: item.productId })),
+      ],
+      []
+    );
   }
 
   async getAllUsers({ filter, pagination, sort }) {
-    let filteredItems = this.filterItems(filter);
-    let aggregatedItems = this.aggregateItems(filteredItems, pagination, sort);
+    const filteredItems = this.filterItems(filter);
+    const aggregatedItems = this.aggregateItems(
+      filteredItems,
+      pagination,
+      sort
+    );
 
     const [users] = await User.aggregate([
       {
@@ -611,8 +622,16 @@ class UserService extends FilterHelper {
   }
 
   async deleteUser(id) {
-    const res = await User.findByIdAndDelete(id).exec();
-    return res || new Error(USER_NOT_FOUND);
+    const user = await User.findById(id).exec();
+
+    if (!user) {
+      throw new RuleError(USER_NOT_FOUND, NOT_FOUND);
+    }
+
+    if (user.role === SUPERADMIN) {
+      throw new RuleError(SUPER_ADMIN_IS_IMMUTABLE, FORBIDDEN);
+    }
+    return User.findByIdAndDelete(id).exec();
   }
 
   async confirmUser(token) {
@@ -767,7 +786,7 @@ class UserService extends FilterHelper {
       user._id,
       {
         $set: {
-          otp_code: otp_code,
+          otp_code,
         },
       },
       { new: true }
